requirements.txt
beautifulsoup4==4.14.3
keras==3.12.0
lxml==6.0.2
matplotlib==3.10.7
numpy==2.2.6
pillow==12.0.0
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
tensorflow==2.20.0
